storage: move API related unit test into nydus-api
Move API related unit test into nydus-api.
